Fix improper reading of Lua 5.4 strings
We were not updating the correct size of the string, in our test case
the string's actual size is 15, but because we remove the null byte it
should be 14. We were not giving the correct size to the struct, this
resulted in a test failure that I have yet to push the code for.